+++ packages/terra-overlay/tests/wdio/loading-overlay-spec.js
@@ -1,49 +1,49 @@
 Terra.describeViewports('Loading Overlay', ['huge'], () => {
-  describe('Loading Overlay Default', () => {
-    before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/default-loading-overlay'));
+  it('should display default Loading Overlay', () => {
+    browser.url('/raw/tests/terra-overlay/overlay/loading-overlay/default-loading-overlay');
 
-    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
+    Terra.validates.element('default', { selector: '#terra-LoadingOverlay' });
   });
 
-  describe('Loading Overlay Container', () => {
-    before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/container-loading-overlay'));
+  it('should display Loading Overlay container', () => {
+    browser.url('/raw/tests/terra-overlay/overlay/loading-overlay/container-loading-overlay');
 
-    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
+    Terra.validates.element('container', { selector: '#terra-LoadingOverlay' });
   });
 
-  describe('Loading Custom Message', () => {
-    before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/custom-message-loading-overlay'));
+  it('should display Loading Overlay with custom message', () => {
+    browser.url('/raw/tests/terra-overlay/overlay/loading-overlay/custom-message-loading-overlay');
 
-    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
+    Terra.validates.element('custom message', { selector: '#terra-LoadingOverlay' });
 
     it('Expect default to have full screen', () => {
       expect(browser.getAttribute('#terra-LoadingOverlay', 'class')).contains('fullscreen');
     });
   });
 
-  describe('Animated', () => {
-    before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/animated-loading-overlay'));
+  it('should display animated Loading Overlay', () => {
+    browser.url('/raw/tests/terra-overlay/overlay/loading-overlay/animated-loading-overlay');
 
     it('Expect Icon to be animated', () => {
       expect(browser.getAttribute('#terra-LoadingOverlay div svg', 'class')).contains('is-spin');
     });
   });
 
-  describe('Light Loading Theme', () => {
-    before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/light-loading-overlay'));
+  it('should display Loading Overlay with light theme', () => {
+    browser.url('/raw/tests/terra-overlay/overlay/loading-overlay/light-loading-overlay');
 
-    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
+    Terra.validates.element('light theme', { selector: '#terra-LoadingOverlay' });
   });
 
-  describe('Dark Loading Overlay Theme', () => {
-    before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/dark-loading-overlay'));
+  it('should display Loading Overlay with dark theme', () => {
+    browser.url('/raw/tests/terra-overlay/overlay/loading-overlay/dark-loading-overlay');
 
-    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
+    Terra.validates.element('dark theme', { selector: '#terra-LoadingOverlay' });
   });
 
-  describe('Clear Loading Overlay Theme', () => {
-    before(() => browser.url('/#/raw/tests/terra-overlay/overlay/loading-overlay/clear-loading-overlay'));
+  it('should display Loading Overlay with clear theme', () => {
+    browser.url('/raw/tests/terra-overlay/overlay/loading-overlay/clear-loading-overlay');
 
-    Terra.it.validatesElement({ selector: '#terra-LoadingOverlay' });
+    Terra.validates.element('clear theme', { selector: '#terra-LoadingOverlay' });
   });
 });