+++ packages/terra-signature/tests/wdio/signature-spec.js
@@ -3,40 +3,34 @@ Terra.describeViewports('Signature', ['medium'], () => {
     browser.url('/#/raw/tests/terra-signature/signature/signature-default');
   });
 
-  describe('Default', () => {
-    Terra.it.validatesElement();
+  it('should display default Signature', () => {
+    Terra.validates.element('default');
   });
 
-  describe('Right Click does not draw lines', () => {
-    it('tries to draw with right click', () => {
-      browser.moveToObject('#drawline', 0, 0);
-      browser.buttonDown(2);
-      browser.moveToObject('#drawline', 90, 90);
-      browser.buttonUp(2);
-    });
+  it('should try to draw with right click', () => {
+    browser.moveToObject('#drawline', 0, 0);
+    browser.buttonDown(2);
+    browser.moveToObject('#drawline', 90, 90);
+    browser.buttonUp(2);
 
-    Terra.it.validatesElement();
+    Terra.validates.element('draw with right click');
   });
 
-  describe('Middle Click does not draw lines', () => {
-    it('tries to draw with middle click', () => {
-      browser.moveToObject('#drawline', 0, 0);
-      browser.buttonDown(1);
-      browser.moveToObject('#drawline', 90, 90);
-      browser.buttonUp(1);
-    });
+  it('should try to draw with middle click', () => {
+    browser.moveToObject('#drawline', 0, 0);
+    browser.buttonDown(1);
+    browser.moveToObject('#drawline', 90, 90);
+    browser.buttonUp(1);
 
-    Terra.it.validatesElement();
+    Terra.validates.element('draw with middle click');
   });
 
-  describe('Drawing Lines', () => {
-    it('draws lines', () => {
-      browser.moveToObject('#drawline', 0, 0);
-      browser.buttonDown(0);
-      browser.moveToObject('#drawline', 90, 90);
-      browser.buttonUp(0);
-    });
+  it('should draws lines', () => {
+    browser.moveToObject('#drawline', 0, 0);
+    browser.buttonDown(0);
+    browser.moveToObject('#drawline', 90, 90);
+    browser.buttonUp(0);
 
-    Terra.it.validatesElement();
+    Terra.validates.element('draws lines');
   });
 });