+++ packages/terra-progress-bar/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Update wdio tests to remove deprecated Terra.it helpers
+
 ## 4.28.0 - (August 11, 2020)
 
 * Changed